MICROSCOPIC bugfix in AuthComponent related to /back; misc. changes
The code added to setCancelledRoute() is meant to solve the problem of using /back to go to an Auth page that redirected from a cancelled route. Nothing was functionally wrong, but the way I coded it meant that even though you didn't navigate from a cancelled route using /back, it still displayed the error message on Auth as if you did.